+++ docs/ReleaseNotes.md
@@ -1,6 +1,21 @@
 Release Notes for Version 14
 ============================
 
+Build 014
+-------
+Published as version 14.9.0
+
+New Features:
+* Add ruby & template style escaping for pseudo localization.
+    * Here is an example of `%RUBY%` substitution parameters.
+    * Here is an example of template substitution parameters:
+    `<% code here %>`
+    * Also added Objective-C/Swift substitution parameter support
+      for the "c" style of escaping
+
+Bug Fixes:
+* Updated `iddarea.json`, `phoneloc.json`, and `numplan.json` from phonenumber library based on metadata version 4.0.0 which are used in PhoneNumber formatting.
+
 Build 013
 -------
 Published as version 14.8.0
@@ -10,7 +25,7 @@ New Features:
   script, region) are valid ISO codes.
 * Supported additional mt-MT, tg-TJ, tk-TM, wo-SN, zu-ZA locales
     * validated locale data and added many test cases
-    
+
 Bug Fixes:
 * Fixed a bug which a default script for `tk` should be `Latin` instead of `Arabic`
 * Fixed a bug which a default script for `tg` should be `Cyrl` instead of `Arabic`